# bff/generate/server/server.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.0 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/go-chi/chi/v5"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// daprSubscribe
	// (GET /dapr/subscribe)
	DaprSubscribe(w http.ResponseWriter, r *http.Request)
	// Find payments
	// (GET /payment/)
	Find(w http.ResponseWriter, r *http.Request, params FindParams)
	// Update Payment
	// (PUT /payment/)
	Update(w http.ResponseWriter, r *http.Request)
	// Create Payment for SAGA
	// (POST /payment/create)
	Create(w http.ResponseWriter, r *http.Request)
	// Reject Create Payment for SAGA
	// (POST /payment/rejectCreate)
	RejectCreate(w http.ResponseWriter, r *http.Request)
	// Delete payment by paymentNo
	// (DELETE /payment/{paymentNo})
	Delete(w http.ResponseWriter, r *http.Request, paymentNo string)
	// Find order by PaymentNo
	// (GET /payment/{paymentNo})
	GetByPaymentNo(w http.ResponseWriter, r *http.Request, paymentNo string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.HandlerFunc) http.HandlerFunc

// DaprSubscribe operation middleware
func (siw *ServerInterfaceWrapper) DaprSubscribe(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DaprSubscribe(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Find operation middleware
func (siw *ServerInterfaceWrapper) Find(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params FindParams

	// ------------- Optional query parameter "customerId" -------------
	if paramValue := r.URL.Query().Get("customerId"); paramValue != "" {

	}

	err = runtime.BindQueryParameter("form", true, false, "customerId", r.URL.Query(), &params.CustomerId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "customerId", Err: err})
		return
	}

	// ------------- Optional query parameter "orderDateFrom" -------------
	if paramValue := r.URL.Query().Get("orderDateFrom"); paramValue != "" {

	}

	err = runtime.BindQueryParameter("form", true, false, "orderDateFrom", r.URL.Query(), &params.OrderDateFrom)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orderDateFrom", Err: err})
		return
	}

	// ------------- Optional query parameter "orderDateTo" -------------
	if paramValue := r.URL.Query().Get("orderDateTo"); paramValue != "" {

	}

	err = runtime.BindQueryParameter("form", true, false, "orderDateTo", r.URL.Query(), &params.OrderDateTo)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orderDateTo", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Find(w, r, params)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Update operation middleware
func (siw *ServerInterfaceWrapper) Update(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Update(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Create operation middleware
func (siw *ServerInterfaceWrapper) Create(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Create(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// RejectCreate operation middleware
func (siw *ServerInterfaceWrapper) RejectCreate(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RejectCreate(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Delete operation middleware
func (siw *ServerInterfaceWrapper) Delete(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "paymentNo" -------------
	var paymentNo string

	err = runtime.BindStyledParameter("simple", false, "paymentNo", chi.URLParam(r, "paymentNo"), &paymentNo)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentNo", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Delete(w, r, paymentNo)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetByPaymentNo operation middleware
func (siw *ServerInterfaceWrapper) GetByPaymentNo(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "paymentNo" -------------
	var paymentNo string

	err = runtime.BindStyledParameter("simple", false, "paymentNo", chi.URLParam(r, "paymentNo"), &paymentNo)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "paymentNo", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetByPaymentNo(w, r, paymentNo)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/dapr/subscribe", wrapper.DaprSubscribe)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/payment/", wrapper.Find)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/payment/", wrapper.Update)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payment/create", wrapper.Create)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/payment/rejectCreate", wrapper.RejectCreate)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/payment/{paymentNo}", wrapper.Delete)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/payment/{paymentNo}", wrapper.GetByPaymentNo)
	})

	return r
}